Conflicts:
	copasi/CopasiDataModel/CCopasiDataModel.cpp
	copasi/UI/CQExperimentData.cpp
	copasi/UI/CQTaskMethodWidget.cpp
	copasi/bindings/perl/local.cpp
	copasi/elementaryFluxModes/CEFMAlgorithm.cpp
	copasi/function/CEvaluationLexer_lex.cpp
	copasi/lna/CLNAMethod.cpp
	copasi/lna/CLNATask.cpp
	copasi/model/CChemEqInterface.cpp
	copasi/model/CModel.cpp
	copasi/model/CProcessQueue.cpp
	copasi/parameterFitting/CExperiment.cpp
	copasi/parameterFitting/CExperiment.h
	copasi/parameterFitting/CExperimentSet.cpp
	copasi/parameterFitting/CExperimentSet.h
	copasi/parameterFitting/CFitProblem.cpp
	copasi/sedml/CSEDMLExporter.cpp
	copasi/steadystate/CNewtonMethod.cpp
	copasi/trajectory/CHybridMethodODE45.cpp
	copasi/trajectory/CTrajectoryTask.cpp
	copasi/trajectory/CTrajectoryTask.h
	copasi/tssanalysis/CTSSAMethod.cpp
	copasi/utilities/CUnit.cpp
	copasi/xml/CCopasiXMLParser.cpp

+++ copasi/function/CEvaluationLexer.h
@@ -1,22 +1,14 @@
-// Begin CVS Header
-//   $Source: /Volumes/Home/Users/shoops/cvs/copasi_dev/copasi/function/CEvaluationLexer.h,v $
-//   $Revision: 1.15 $
-//   $Name:  $
-//   $Author: shoops $
-//   $Date: 2011/05/26 12:25:40 $
-// End CVS Header
-
-// Copyright (C) 2011 - 2010 by Pedro Mendes, Virginia Tech Intellectual
+// Copyright (C) 2010 - 2015 by Pedro Mendes, Virginia Tech Intellectual
 // Properties, Inc., University of Heidelberg, and The University
 // of Manchester.
 // All rights reserved.
 
-// Copyright (C) 2008 by Pedro Mendes, Virginia Tech Intellectual
+// Copyright (C) 2008 - 2009 by Pedro Mendes, Virginia Tech Intellectual
 // Properties, Inc., EML Research, gGmbH, University of Heidelberg,
 // and The University of Manchester.
 // All rights reserved.
 
-// Copyright (C) 2001 - 2007 by Pedro Mendes, Virginia Tech Intellectual
+// Copyright (C) 2005 - 2007 by Pedro Mendes, Virginia Tech Intellectual
 // Properties, Inc. and EML Research, gGmbH.
 // All rights reserved.
 
@@ -51,11 +43,11 @@ public:
    * Default constructor
    */
   yyYaccParser():
-      mpNode(NULL),
-      mpRootNode(NULL),
-      mBoolean(false),
-      mpNodeList(NULL),
-      mPosition(0)
+    mpNode(NULL),
+    mpRootNode(NULL),
+    mBoolean(false),
+    mpNodeList(NULL),
+    mPosition(0)
   {}
 
   /**
@@ -152,8 +144,8 @@ public:
   virtual int yywrap() {return 1;}
 
 protected:
-  virtual int LexerInput(char* buf, int max_size);
-  virtual void LexerOutput(const char* buf, int size);
+  virtual size_t LexerInput(char* buf, size_t max_size);
+  virtual void LexerOutput(const char* buf, size_t size);
   virtual void LexerError(const char* msg);
 
   void yyunput(int c, char* buf_ptr);
